methods moved out of navbarPage
