+++ src/test/java/features/bookAppointment.feature
@@ -23,9 +23,9 @@ Feature: End to end functions of booking and viewing Appointments in patient por
     Examples:
       | Payer type | Appointment type |
       | Patient    | Consultation     |
-#      | Insurer    | Consultation     |
-#      | Employer   | Consultation     |
-#      | Patient    | Health Screen    |
+      | Insurer    | Consultation     |
+      | Employer   | Consultation     |
+      | Patient    | Health Screen    |
 
 
   @BookByAppointmentFilters
@@ -50,9 +50,9 @@ Feature: End to end functions of booking and viewing Appointments in patient por
 
     Examples:
       | Payer type | Appointment type | Clinician     | Site          | Time      | Date       |
-      | Patient    | Consultation     | Any Clinician | Work location | Afternoon | 3/05/2020 |
-      | Insurer    | Consultation     | Kaaru kaaru   | Work Location | Morning   | 07/05/2020 |
-      | Insurer    | Consultation     | Kaaru kaaru   | Home          | Morning   | 02/05/2020 |
+      | Patient    | Consultation     | Any Clinician | Work location | Afternoon | 20/05/2020 |
+      | Insurer    | Consultation     | Kaaru kaaru   | Work Location | Morning   | 20/05/2020 |
+      | Insurer    | Consultation     | Kaaru kaaru   | Home          | Morning   | 20/05/2020 |
 
 
   @viewExistingAppointments
