+++ src/test/java/runner/RunCukeTestng.java
@@ -28,10 +28,10 @@ import utils.utilities;
 
 @CucumberOptions(
 
-       // features = {"src/test/java/features/bookAppointment.feature"},
+       features = {"src/test/java/features/bookAppointment.feature"},
         glue = {"stepDefinitions"},
         monochrome = true,
-      //  tags = {"@smoke"},
+        tags = {"@BookByAppointmentFilters"},
