+++ src/test/java/stepDefinitions/logintoapplicationsteps.java
@@ -34,7 +34,9 @@ public class logintoapplicationsteps extends baseStepDefs {
 
 	@Then("^I should be directed to home page$")
 	public void i_should_be_directed_to_home_page() throws Throwable {
-	    navbarpage.VerifyBookAppointment();
+
+		bookappointment.VerifyBookAppointment();
+	    //navbarpage.VerifyBookAppointment();
 	}
